混乱代码，需要重构

+++ Project-HEW2020/debug_font.cpp
@@ -26,7 +26,7 @@ static Sprite* g_pSpriteDebug = NULL;
 void DebugFont_Init(void)
 {
     g_pSpriteDebug = new SpriteNormal(TEXTURE_DEBUG);
-    g_pSpriteDebug->SetCutRange(DEBUG_FONT_WIDTH, DEBUG_FONT_HEIGHT);
+    g_pSpriteDebug->SetCutRange((float)DEBUG_FONT_WIDTH, (float)DEBUG_FONT_HEIGHT);
     g_pSpriteDebug->SetPolygonSize(32.0f, 64.0f);
     g_pSpriteDebug->SetColor(D3DCOLOR_RGBA(155, 155, 255, 255));
 }
@@ -43,11 +43,11 @@ void DebugFont_Draw(float dx, float dy, const char *pString)
     for (int i = 0; i < strlen(pString); i++)
     {
         int offset = pString[i] - ' ';
-        int tcx = offset % DEBUG_FONT_START_CHAR_CODE * DEBUG_FONT_WIDTH;
-        int tcy = offset / DEBUG_FONT_ONE_LINE_CHAR_COUNT * DEBUG_FONT_HEIGHT;
+        float tcx = (float)(offset % DEBUG_FONT_START_CHAR_CODE * DEBUG_FONT_WIDTH);
+        float tcy = (float)(offset / DEBUG_FONT_ONE_LINE_CHAR_COUNT * DEBUG_FONT_HEIGHT);
 
         g_pSpriteDebug->SetCutPos(tcx, tcy);
-        g_pSpriteDebug->SetDrawPos(dx + i * DEBUG_FONT_WIDTH, dy);
+        g_pSpriteDebug->SetDrawPos(dx + i * (float)DEBUG_FONT_WIDTH, dy);
         g_pSpriteDebug->Draw();
     }
 }
